[NUGUDI-66] refactor(web): 컴포넌트 함수형으로 변경 및, 컴포넌트 폴더 구조 통일화
