+++ apps/web/app/_ui/views/email-login-view/index.tsx
@@ -1,6 +1,6 @@
-import WelcomeText from "../components/welcome-text";
-import EmailLoginSection from "../sections/email-login-section";
-import * as styles from "./email-login-view.css";
+import WelcomeText from "../../components/welcome-text";
+import EmailLoginSection from "../../sections/email-login-section";
+import * as styles from "./index.css";
 
 const EmailLoginView = () => {
   return (
